Refactor and add readonly
